implement qr code in spacewalk bridge
