+++ package.json
@@ -54,6 +54,7 @@
     "luxon": "^3.2.1",
     "match-sorter": "^6.3.1",
     "preact": "^10.12.1",
+    "qrcode.react": "^3.1.0",
     "react-daisyui": "^3.1.2",
     "react-device-detect": "^2.2.3",
     "react-hook-form": "^7.43.2",
